close over checks and parallelize with a waitgroup
