+++ cla.go
@@ -13,7 +13,7 @@ import (
 	"fmt"
 	"net/http"
 	"regexp"
-	"strings"
+	"sync"
 
 	"github.com/google/go-github/v38/github"
 )
@@ -37,6 +37,8 @@ type Details struct {
 var stringMatchers = []string{"\bCLA\b", "Contributor License Agreement"}
 var actionMatcher = "uses:[[:space:]]*?cla-assistant/github-action"
 var prLabelMatcher = "cla:[[:space:]]*?[yes|no]"
+
+var tree *github.Tree
 
 func (d *Details) Required() bool {
 	if d.Tag || d.BotFile || d.InContributing || d.InREADME || d.Action {
@@ -66,39 +68,54 @@ func DetailWithContext(ctx context.Context, client *github.Client, owner string,
 	if resp.StatusCode == http.StatusUnauthorized || resp.StatusCode == http.StatusForbidden {
 		return &Details{}, ErrInvalidToken
 	}
-	known := isKnown(owner)
-
 	def := r.GetDefaultBranch()
 
-	// TODO: parallelize
-	tag, tagerr := hasCLATag(ctx, client, owner, repo)
-	bot, boterr := hasCLABotFile(ctx, client, owner, repo, def)
-	contrib, contriberr := referencesCLAInContributing(ctx, client, owner, repo, def)
-	readme, readmeerr := referencesCLAInREADME(ctx, client, owner, repo, def)
-	act, acterr := usesCLAAssistantAction(ctx, client, owner, repo, def)
-
-	d := &Details{
-		Known:          known,
-		Tag:            tag,
-		BotFile:        bot,
-		InContributing: contrib,
-		InREADME:       readme,
-		Action:         act,
+	type check = func(wg *sync.WaitGroup, d *Details, e *Errors)
+	checks := []check{
+		func(wg *sync.WaitGroup, d *Details, e *Errors) {
+			defer wg.Done()
+			d.Known = isKnown(owner)
+		},
+		func(wg *sync.WaitGroup, d *Details, e *Errors) {
+			defer wg.Done()
+			d.Tag, e.TagErr = hasCLATag(ctx, client, owner, repo)
+		},
+		func(wg *sync.WaitGroup, d *Details, e *Errors) {
+			defer wg.Done()
+			d.BotFile, e.BotFileErr = hasCLABotFile(ctx, client, owner, repo, def)
+		},
+		func(wg *sync.WaitGroup, d *Details, e *Errors) {
+			defer wg.Done()
+			d.InContributing, e.InContributingErr = referencesCLAInContributing(ctx, client, owner, repo, def)
+		},
+		func(wg *sync.WaitGroup, d *Details, e *Errors) {
+			defer wg.Done()
+			d.InREADME, e.InREADMEErr = referencesCLAInREADME(ctx, client, owner, repo, def)
+		},
+		func(wg *sync.WaitGroup, d *Details, e *Errors) {
+			defer wg.Done()
+			d.Action, e.ActionErr = usesCLAAssistantAction(ctx, client, owner, repo, def)
+		},
 	}
+	var wg sync.WaitGroup
+	wg.Add(len(checks))
 
-	e := &Errors{
-		TagErr:            tagerr,
-		BotFileErr:        boterr,
-		InContributingErr: contriberr,
-		InREADMEErr:       readmeerr,
-		ActionErr:         acterr,
+	var (
+		d Details
+		e Errors
+	)
+	for _, c := range checks {
+		// c(&wg, &d, &e)
+		go c(&wg, &d, &e)
 	}
+	wg.Wait()
 
-	return d, e.ErrOrNil()
+	return &d, e.ErrOrNil()
 }
 
 func find(ctx context.Context, client *github.Client, owner, repo, branch, path string) (*github.TreeEntry, error) {
 	tree, _, err := client.Git.GetTree(ctx, owner, repo, branch, true)
+	
 	if err != nil {
 		return nil, fmt.Errorf("failed to get %s/%s tree: %v", owner, repo, err)
 	}
@@ -152,73 +169,6 @@ func contentAtSHA(ctx context.Context, client *github.Client, owner, repo, sha s
 	return base64.StdEncoding.DecodeString(b.GetContent())
 }
 
-func isKnown(owner string) bool {
-	for _, o := range knownOwners {
-		if o == owner {
-			return true
-		}
-	}
-	return false
-}
-
-func hasCLATag(ctx context.Context, client *github.Client, owner, repo string) (bool, error) {
-	opts := &github.PullRequestListOptions{
-		ListOptions: github.ListOptions{
-			PerPage: 100,
-		},
-		State: "all",
-	}
-	prs, _, err := client.PullRequests.List(ctx, owner, repo, opts)
-	if err != nil {
-		return false, fmt.Errorf("error getting %s/%s PRs: %v", owner, repo, err)
-	}
-	errs := make([]error, 0, 100)
-	for _, pr := range prs {
-		for _, label := range pr.Labels {
-			match, err := regexp.Match(prLabelMatcher, []byte(label.GetName()))
-			if err != nil {
-				errs = append(errs, err)
-			}
-			if match {
-				return true, nil
-			}
-		}
-	}
-	if len(errs) != 0 {
-		var lines []string
-		for _, err := range errs {
-			lines = append(lines, fmt.Sprintf("* %v", err))
-		}
-		return false, fmt.Errorf("%d errors(s) checking recent PR labels:\n\t%s", len(errs), strings.Join(lines, "\n\t"))
-	}
-
-	return false, nil
-}
-
-func hasCLABotFile(ctx context.Context, client *github.Client, owner, repo, branch string) (bool, error) {
-	te, err := find(ctx, client, owner, repo, branch, ".clabot")
-	if te == nil || err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-func referencesCLAInContributing(ctx context.Context, client *github.Client, owner, repo, branch string) (bool, error) {
-	content, err := contentAtPath(ctx, client, owner, repo, branch, "CONTRIBUTING.md")
-	if err != nil {
-		return false, fmt.Errorf("failed to check CONTRIBUTING.md: %v", err)
-	}
-	return referencesCLAInContent(content)
-}
-
-func referencesCLAInREADME(ctx context.Context, client *github.Client, owner, repo, branch string) (bool, error) {
-	content, err := contentAtPath(ctx, client, owner, repo, branch, "README.md")
-	if err != nil {
-		return false, fmt.Errorf("failed to check README.md: %v", err)
-	}
-	return referencesCLAInContent(content)
-}
-
 func referencesCLAInContent(content []byte) (bool, error) {
 	var match bool
 	var err error
@@ -230,47 +180,6 @@ func referencesCLAInContent(content []byte) (bool, error) {
 		if match {
 			return true, nil
 		}
-	}
-
-	return false, nil
-}
-
-func usesCLAAssistantAction(ctx context.Context, client *github.Client, owner, repo, branch string) (bool, error) {
-	workflowsEntry, err := find(ctx, client, owner, repo, branch, ".github/workflows")
-	if err != nil {
-		if err == ErrTruncatedTree {
-			return false, fmt.Errorf("tree was truncated and .github/workflows was possibly missed")
-		}
-		return false, err
-	}
-	workflowsTree, _, err := client.Git.GetTree(ctx, owner, repo, workflowsEntry.GetSHA(), false)
-	if err != nil {
-		return false, fmt.Errorf("failed to get %s/%s/master/.github/workflows tree: %v", owner, repo, err)
-	}
-
-	errs := make(map[string]error)
-	for _, e := range workflowsTree.Entries {
-		content, err := contentAtSHA(ctx, client, owner, repo, e.GetSHA())
-		if err != nil {
-			errs[e.GetPath()] = err
-			continue
-		}
-		match, err := regexp.Match(actionMatcher, content)
-		if err != nil {
-			errs[e.GetPath()] = err
-			continue
-		}
-		if match {
-			return true, nil
-		}
-	}
-
-	if len(errs) != 0 {
-		var lines []string
-		for path, err := range errs {
-			lines = append(lines, fmt.Sprintf("* %s: %v", path, err))
-		}
-		return false, fmt.Errorf("%d error(s) checking for cla-assistant action:\n\t%s", len(errs), strings.Join(lines, "\n\t"))
 	}
 
 	return false, nil
